# docs/swagger.yaml
definitions:
...
  dbmodels.SalesType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Acquitisions:
    properties:
      address:
        type: string
      best_visit:
        type: string
      business_location:
        type: string
      business_type:
        type: string
      category_type:
        type: string
      city_id:
        type: string
      company_code:
        type: string
      customer_code:
        type: string
      device:
        type: integer
      district_id:
        type: string
      dob:
        type: string
      education:
        type: boolean
      id_card:
        type: string
      image_id_card:
        description: images
        type: string
      image_merchant:
        description: images
        type: string
      image_merchant_location:
        description: images
        type: string
      image_merchant_location_additional_1:
        type: string
      image_merchant_location_additional_2:
        type: string
      institution:
        type: string
      latitude:
        type: string
      lokasi_usaha:
        type: string
      longitude:
        type: string
      merchant_id:
        type: string
      mpan:
        type: string
      name:
        type: string
      nmid:
        type: string
      note:
        type: string
      operation_hour:
        type: string
      owner_name:
        type: string
      phone_number:
        type: string
      photo_location_left:
        type: string
      photo_location_right:
        type: string
      photo_merchant_location:
        description: images
        type: string
      photo_qr_preprinted:
        type: string
      province_id:
        type: string
      sales;omitempty:
        type: boolean
      sales_id:
        type: integer
      sales_name:
        type: string
      sales_phone:
        type: string
      signature:
        description: images
        type: string
      store_name_preprinted:
        type: string
      version_id:
        type: integer
      village_id:
        type: string
    type: object
  models.AttendancePostReq:
    properties:
      attendance_category:
        type: string
      attendance_category_type:
        type: string
      latitude:
        type: string
      location:
        type: string
      longitude:
        type: string
      notes:
        type: string
      photo_accuration:
        type: string
      photo_selfie:
        type: string
      sales_id:
        type: integer
      sales_name:
        type: string
      sales_phone:
        type: string
      time:
        type: string
      type_attendance:
        type: string
      type_timezone:
        type: string
    type: object
  models.BusinessTypeListBySRReq:
    properties:
      sales_type_id:
        items:
          type: integer
        type: array
    type: object
  models.BusinessTypeListRes:
    properties:
      tipe_bisnis:
        items:
          $ref: '#/definitions/models.DataModule'
        type: array
    type: object
...
  models.ChangePhotoPinKTPRes:
    properties:
      description_code:
        type: string
      response_code:
        type: string
    type: object
...
  models.DataModule:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
...
  models.LoginReq:
    properties:
      deviceId:
        type: string
      deviceToken:
        type: string
      firebase_token:
        type: string
      phoneNumber:
        type: string
      pin:
        type: string
      role:
        type: string
      salesId:
        type: string
      version_code:
        type: string
    type: object
  models.LoginRes:
    properties:
      description_code:
        type: string
      email:
        type: string
      force_update:
        type: boolean
      functional_id:
        type: string
      phone:
        type: string
      photo:
        type: string
      response_code:
        type: string
      sales_id:
        type: string
      sales_name:
        type: string
      sales_type:
        items:
          $ref: '#/definitions/dbmodels.SalesType'
        type: array
      session_token:
        type: string
      sfa_id:
        type: string
      status:
        type: string
    type: object
  models.LoginV23Res:
    properties:
      description_code:
        type: string
      email:
        type: string
      force_update:
        type: boolean
      functional_id:
        type: string
      phone:
        type: string
      photo:
        type: string
      response_code:
        type: string
      sales_id:
        type: string
      sales_name:
        type: string
      sales_type_id:
        type: integer
      session_token:
        type: string
      sfa_id:
        type: string
      status:
        type: string
    type: object
...
  models.MetaData:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Ok
        type: string
      status:
        example: true
        type: boolean
    type: object
...
  models.Response:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/models.MetaData'
        type: object
    type: object
...
  models.TodolistDetailReq:
    properties:
      customer_code:
        example: "543543422556.0"
        type: string
      merchant_phone:
        example: "089898988897"
        type: string
      todolist_category_id:
        example: 5
        type: integer
      todolist_id:
        example: "322"
        type: string
    type: object
...
  models.TodolistPostReq:
    properties:
      alamat:
        type: string
      body:
        items:
          type: string
        type: array
      city:
        type: string
      content_type:
        items:
          type: string
        type: array
      district:
        type: string
      label:
        items:
          type: string
        type: array
      label_photo_merchant1:
        type: string
      label_photo_merchant2:
        type: string
      label_photo_merchant3:
        type: string
      latitude:
        type: string
      latitude_merchant:
        type: string
      longitude:
        type: string
      longitude_merchant:
        type: string
      new_task_date:
        type: string
      old_task_date:
        type: string
      patokan:
        type: string
      photo_merchant1:
        type: string
      photo_merchant2:
        type: string
      photo_merchant3:
        type: string
      province:
        type: string
      reason:
        type: string
      status:
        type: string
      task_id:
        items:
          type: integer
        type: array
      todolist_id:
        type: integer
      versionId:
        type: integer
      village:
        type: string
    type: object
  models.TodolistPostV23Req:
    properties:
      acquitisions:
        $ref: '#/definitions/models.Acquitisions'
        type: object
      alamat:
        type: string
      body:
        items:
          type: string
        type: array
      city:
        type: string
      content_type:
        items:
          type: string
        type: array
      district:
        type: string
      id_merchant:
        type: integer
      label:
        items:
          type: string
        type: array
      label_photo_merchant1:
        type: string
      label_photo_merchant2:
        type: string
      label_photo_merchant3:
        type: string
      latitude:
        type: string
      latitude_merchant:
        type: string
      longitude:
        type: string
      longitude_merchant:
        type: string
      new_task_date:
        type: string
      old_task_date:
        type: string
      patokan:
        type: string
      photo_merchant1:
        type: string
      photo_merchant2:
        type: string
      photo_merchant3:
        type: string
      province:
        type: string
      reason:
        type: string
      status:
        type: string
      task_id:
        items:
          type: integer
        type: array
      todolist_category_id:
        type: string
      todolist_id:
        type: integer
      versionId:
        type: integer
      village:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v2.2/sfa/attendance/post:
    post:
      consumes:
      - application/json
      description: Attendance Post
      operationId: Attendance Post v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AttendancePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Attendance Post
      tags:
      - OTTO SFA
...
  /v2.2/sfa/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginRes'
              type: object
      summary: Login
      tags:
      - OTTO SFA
...
  /v2.2/sfa/todolist/post:
    post:
      consumes:
      - application/json
      description: Todolist Post
      operationId: Todolist Post v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TodolistPostReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            $ref: '#/definitions/models.Response'
      summary: Todolist Post
      tags:
      - OTTO SFA
...
  /v2.3/merchants/business_type/list:
    get:
      consumes:
      - application/json
      description: Business Type List
      operationId: Business Type List v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Business Type List Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BusinessTypeListRes'
              type: object
      summary: Business Type List
      tags:
      - OTTO SFA
    post:
      consumes:
      - application/json
      description: Business Type List  By SR
      operationId: Business Type List  By SR
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.BusinessTypeListBySRReq'
      produces:
      - application/json
      responses:
        "200":
          description: Business Type List By SR Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.BusinessTypeListRes'
              type: object
      summary: Business Type List  By SR
      tags:
      - OTTO SFA
...
  /v2.3/sfa-idm/todolist/detail:
    post:
      consumes:
      - application/json
      description: Todolist Detail
      operationId: Todolist Detail IDM
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TodolistDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Detail Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Todolist Detail
      tags:
      - OTTO SFA-IDM
...
  /v2.3/sfa/attendance/post:
    post:
      consumes:
      - application/json
      description: Attendance Post
      operationId: Attendance Post v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AttendancePostReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Attendance Post
      tags:
      - OTTO SFA
  /v2.3/sfa/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginV23Res'
              type: object
      summary: Login
      tags:
      - OTTO SFA
...
  /v2.3/sfa/todolist/detail:
    post:
      consumes:
      - application/json
      description: Todolist Detail
      operationId: Todolist Detail v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TodolistDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Detail Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Todolist Detail
      tags:
      - OTTO SFA
...
  /v2.3/sfa/todolist/post:
    post:
      consumes:
      - application/json
      description: Todolist Post
      operationId: Todolist Post v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TodolistPostV23Req'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            $ref: '#/definitions/models.Response'
      summary: Todolist Post
      tags:
      - OTTO SFA
...
  /v2.4/sfa/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login v2.4
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginV23Res'
              type: object
      summary: Login
      tags:
      - OTTO SFA
...
  /v2.4/sfa/todolist/detail:
    post:
      consumes:
      - application/json
      description: Todolist Detail
      operationId: Todolist Detail v2.4
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TodolistDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Detail Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Todolist Detail
      tags:
      - OTTO SFA
...
  /v2.4/sfa/todolist/post:
    post:
      consumes:
      - application/json
      description: Todolist Post
      operationId: Todolist Post v2.4
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.TodolistPostV23Req'
      produces:
      - application/json
      responses:
        "200":
          description: Todolist Post Response EXAMPLE
          schema:
            $ref: '#/definitions/models.Response'
      summary: Todolist Post
      tags:
      - OTTO SFA
...
swagger: "2.0"